DEV: Add a shared locations extension API
